+++ src/plugins/geofence.ts
@@ -1,5 +1,5 @@
 import { Cordova, Plugin } from './plugin';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from '@reactivex/rxjs';
 /**
  * @name Geofence
  * @description Monitors circular geofences around latitude/longitude coordinates, and sends a notification to the user when the boundary of a geofence is crossed. Notifications can be sent when the user enters and/or exits a geofence.
@@ -96,7 +96,7 @@ export class Geofence {
    * @return {Promise<any>}
    */
   @Cordova()
-  static initialize(): Promise<void> { return };
+  static initialize(): Promise<void> { return; };
 
   /**
    * Adds a new geofence or array of geofences. For geofence object, see above.
@@ -104,7 +104,7 @@ export class Geofence {
    * @return {Promise<any>}
    */
   @Cordova()
-  static addOrUpdate(geofences: Object | Array<Object>): Promise<void> { return };
+  static addOrUpdate(geofences: Object | Array<Object>): Promise<void> { return; };
 
   /**
    * Removes a geofence or array of geofences. `geofenceID` corresponds to one or more IDs specified when the
@@ -113,7 +113,7 @@ export class Geofence {
    * @return {Promise<any>}
    */
   @Cordova()
-  static remove(geofenceId: string | Array<string>): Promise<void> { return };
+  static remove(geofenceId: string | Array<string>): Promise<void> { return; };
 
   /**
    * Removes all geofences.
@@ -121,7 +121,7 @@ export class Geofence {
    * @return {Promise<any>}
    */
   @Cordova()
-  static removeAll(): Promise<void> { return };
+  static removeAll(): Promise<void> { return; };
 
   /**
    * Returns an array of geofences currently being monitored.
@@ -129,11 +129,10 @@ export class Geofence {
    * @return {Promise<Array<string>>}
    */
   @Cordova()
-  static getWatched(): Promise<string> { return };
+  static getWatched(): Promise<string> { return; };
 
   /**
    * Called when a geofence is crossed in the direction specified by `TransitType`.
-   * Commenting out. Not yet implemented in plugin.
    *
    * @return {Promise<any>}
    */
@@ -147,8 +146,7 @@ export class Geofence {
   }
 
   /**
-   * Called when the user clicks a geofence notification. iOS and Android only.
-   * Commenting out. Not yet implemented in plugin.
+   * Called when the user clicks a geofence notification. iOS and Android only.   
    *
    * @return {Promise<Object>}
    */